Basic explanation of the TCP/IP interface to the server written
This commit only has to do with the manual, so no source code has been edited, only the TeX file (and the corresponding PDF). I explain in broad terms what the TCP/IP communication interface is, in terms of how JSON-RPC commands have to be sent. Not all commands have been documented yet, but the general structure should be clear