Wrote the basic JSON-RPC call interface for the fitter-plotter
The interface is in file JSONinterpreter.py It's one class, called JSONread(). There's also the file test_JSONinterpreter.py, which has the unit tests, written in pytest framework. The old functions are still in the file and possibly the program will not run now, but this is the commit to record the newly-created interface JSON-RPC