JSON-RPC2 interface written
JSON-RPC 2 interface is written and works as intended, at least for the first three of the methods. The rest of the methods has to be still checked and written, but that's only about adding new functionality, rather than any fundamental changes. In fact, it's possible that we will have more than 2 more methods later