Major rewrite finished, plotting and fitting works from JSON-RPC commands
I tested plotting generated datasets, and I tested fitting them with all implemented fitters from scipy.optimize (least-squares, minimize, differential_evolution, shgo, basinhopping, dual_annealing) on an example of a perfect sine wave, and the fit is successful with all fitters. This is of course a test only that the inputs are correct, not that all fit parameters are good, but now when we start working with noisy data, we can see which fitters perform better.

When a fitter doesn't converge and produces succes = False message, we see that in the output.

The transmission of results to the client is not yet implemented