The seach function for minima for resonance tracking is now done.
We do the search basically by minimizing the distance between the absolute value of the signal in resonance tracking, and zero. The fit function is called "resonancetrackingzero". The difficult part is to write down the cost function. Right now it's the absolute square of the difference, with a penalty for short regions that goes with the log of the length and the square root of the length. This is difficult, because it's a highly nonlinear function

Also in this commit there are some bugs corrected, like printout of some debug messages is gone, and in the peak finding routines, we got replaced "smoothing" by "smoothingwidth"