Res tracking zero region finding without error bars
It looks like the error bars in the dataset where one has to find the largest zero region are not helpful. I got rid of them now, and it seems to work better and more robustly. This has to do with the fact that the cost function for the fit is not so easy to estimate