functions clear_everything clear_plot and clear_replot seem to work now
The idea is to create an interface by which the user can either clear everything, which should delete all data that has been sent to the remote fitter, or in the case of clear_plot, to only delete the visual curves, without deleting the unterlying data. It may look weird that there is the clear_replot routine, but it's just there to be able to replot the curve whose visual has been deleted, without resending the data. Since it's so tightly connected to this clear_plot routine, we can say that it should be used together, and that's why we put it here.